added forecast model
