+++ lib/models/forecast_model.dart
@@ -1,0 +1,8 @@
+class ForecastItem {
+  late DateTime date;
+  late Map<String, double> allergenTypeStrength;
+}
+
+class AllergyForecast {
+  late List<ForecastItem> forecast;
+}
